added fix: changed ids data type from Long to int, added additional fk column cart_id to shopping_cart_item table schema

+++ src/main/java/nus/iss/team1/grabfreshfood/model/CartItem.java
@@ -12,10 +12,10 @@ public class CartItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "item_id")
-    private Long cartItemId;
+    private int cartItemId;
 
     @Column(name = "product_id")
-    private Long productId;
+    private int productId;
     private double price;
 
     @Column(name = "is_checkout")
